requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.2
django-cors-headers==4.0.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.8
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
psycopg2-binary==2.9.6
python-decouple==3.8
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
yarl==1.9.2
